+++ src/components/Auth/AuthServiceWrapper.stories.tsx
@@ -1,6 +1,6 @@
 import { ComponentMeta, ComponentStory } from '@storybook/react'
 
-import { authDecorator, authServiceList, WrappedAuthComponent } from '../.storybook'
+import { authDecorator, authServiceList, WrappedAuthComponent } from '../../.storybook'
 import { AuthServiceWrapper } from './AuthServiceWrapper'
 
 const StorybookEntry = {
