+++ src/components/Payload/Details/Details.stories.tsx
@@ -1,6 +1,6 @@
 import { ComponentMeta, ComponentStory } from '@storybook/react'
 
-import { samplePayload } from '../../.storybook'
+import { samplePayload } from '../../../.storybook'
 import { PayloadDetails } from './Details'
 
 const StorybookEntry = {
